Add try/catch to JSON parsing code
I have an API endpoint that returns a status code of 201, but no data when sent a POST. This triggers the error handler (JSON.parse throws an error if it's malformed). 

I believe the success handler should still be triggered since the HTTP request was successful, so this PR swallows any JSON parse errors.